API-12089 copy changes from v3.5
