+++ src/configuration/index.ts
@@ -1,31 +1,32 @@
 import { WebAPI } from "../internal";
 import { TokenGetter } from "../authorization";
 import {
-  AgentFields,
+  AddAutoAccessRequest,
+  AddAutoAccessResponse,
   Agent,
-  EmptyResponse,
-  CreateAgentResponse,
-  BotFields,
-  CreateBotResponse,
-  Bot,
+  AgentFields,
   AgentPriorities,
+  AutoAccess,
+  Bot,
+  BotFields,
+  CreateAgentResponse,
+  CreateBotResponse,
   CreateGroupResponse,
+  EmptyResponse,
+  GetLicenseIDResponse,
+  GetOrganizationIDResponse,
   Group,
-  PropertyConfig,
+  Properties,
   PropertiesConfig,
-  Webhook,
+  PropertyConfig,
   RegisteredWebhook,
   RegisterWebhookResponse,
+  UpdateAutoAccessRequest,
+  WebAPIOptions,
+  Webhook,
   WebhookData,
   WebhooksState,
-  AddAutoAccessResponse,
-  AutoAccess,
-  AddAutoAccessRequest,
-  UpdateAutoAccessRequest,
-  GetOrganizationIDResponse,
-  GetLicenseIDResponse,
 } from "./structures";
-import { Properties, WebAPIOptions } from "../objects/index";
 
 export default class ConfigurationAPI extends WebAPI {
   constructor(clientID: string, tokenGetter: TokenGetter, options?: WebAPIOptions) {
@@ -208,7 +209,7 @@ export default class ConfigurationAPI extends WebAPI {
   /**
    * Returns details about a group specified by its id.
    * @param id - group ID to get info about
-   * @param fields - additiona fields to include
+   * @param fields - additional fields to include
    */
   async getGroup(id: number, fields?: string[]): Promise<Group> {
     return this.send("get_group", { id, fields });
@@ -267,7 +268,7 @@ export default class ConfigurationAPI extends WebAPI {
 
   /**
    * Returns the properties set within a license.
-   * @param namespace_prefix - namsepace prefix
+   * @param namespace_prefix - namespace prefix
    * @param name_prefix - name prefix
    */
   async listLicenseProperties(namespace_prefix?: string, name_prefix?: string): Promise<Properties> {
@@ -296,6 +297,7 @@ export default class ConfigurationAPI extends WebAPI {
       properties,
     });
   }
+
   /**
    * Returns the properties set within a group.
    * @param id - group ID to get properties from
@@ -358,7 +360,7 @@ export default class ConfigurationAPI extends WebAPI {
 
   /**
    * Enables license webhooks for authorization token's clientID
-   * @param clientID - clientID when authorizing via Personal Access Token
+   * @param client_id - clientID when authorizing via Personal Access Token
    */
   async enableLicenseWebhooks(client_id?: string): Promise<EmptyResponse> {
     return this.send("enable_license_webhooks", { client_id });
@@ -366,7 +368,7 @@ export default class ConfigurationAPI extends WebAPI {
 
   /**
    * Disables license webhooks for authorization token's clientID
-   * @param clientID - clientID when authorizing via Personal Access Token
+   * @param client_id - clientID when authorizing via Personal Access Token
    */
   async disableLicenseWebhooks(client_id?: string): Promise<EmptyResponse> {
     return this.send("disable_license_webhooks", { client_id });
@@ -374,7 +376,7 @@ export default class ConfigurationAPI extends WebAPI {
 
   /**
    * Gets license webhooks' state for authorization token's clientID
-   * @param clientID - clientID when authorizing via Personal Access Token
+   * @param client_id - clientID when authorizing via Personal Access Token
    */
   async getLicenseWebhooksState(client_id?: string): Promise<WebhooksState> {
     return this.send("get_license_webhooks_state", { client_id });
